move temp view detection
