+++ spark/v3.5/spark-extensions/src/main/scala/org/apache/spark/sql/catalyst/analysis/CheckViews.scala
@@ -20,9 +20,7 @@
 package org.apache.spark.sql.catalyst.analysis
 
 import org.apache.spark.sql.AnalysisException
-import org.apache.spark.sql.catalyst.expressions.SubqueryExpression
 import org.apache.spark.sql.catalyst.plans.logical.LogicalPlan
-import org.apache.spark.sql.catalyst.plans.logical.View
 import org.apache.spark.sql.catalyst.plans.logical.views.CreateIcebergView
 import org.apache.spark.sql.connector.catalog.Identifier
 import org.apache.spark.sql.connector.catalog.ViewCatalog
@@ -36,7 +34,6 @@ object CheckViews extends (LogicalPlan => Unit) {
       case CreateIcebergView(ResolvedIdentifier(_: ViewCatalog, ident), _, query, columnAliases, _,
       queryColumnNames, _, _, _, _, _) =>
         verifyColumnCount(ident, columnAliases, query)
-        verifyTemporaryObjectsDontExist(ident, query)
         SchemaUtils.checkColumnNameDuplication(queryColumnNames, SQLConf.get.resolver)
 
       case _ => // OK
@@ -61,45 +58,5 @@ object CheckViews extends (LogicalPlan => Unit) {
             "dataColumns" -> query.output.map(c => c.name).mkString(", ")))
       }
     }
-  }
-
-  /**
-   * Permanent views are not allowed to reference temp objects
-   */
-  private def verifyTemporaryObjectsDontExist(
-    name: Identifier,
-    child: LogicalPlan): Unit = {
-    import org.apache.spark.sql.connector.catalog.CatalogV2Implicits._
-
-    val tempViews = collectTemporaryViews(child)
-    tempViews.foreach { nameParts =>
-      throw new AnalysisException(
-        errorClass = "INVALID_TEMP_OBJ_REFERENCE",
-        messageParameters = Map(
-          "obj" -> "VIEW",
-          "objName" -> name.name(),
-          "tempObj" -> "VIEW",
-          "tempObjName" -> nameParts.quoted))
-    }
-
-    // TODO: check for temp function names
-  }
-
-
-  /**
-   * Collect all temporary views and return the identifiers separately
-   */
-  private def collectTemporaryViews(child: LogicalPlan): Seq[Seq[String]] = {
-    def collectTempViews(child: LogicalPlan): Seq[Seq[String]] = {
-      child.flatMap {
-        case view: View if view.isTempView => Seq(view.desc.identifier.nameParts)
-        case plan => plan.expressions.flatMap(_.flatMap {
-          case e: SubqueryExpression => collectTempViews(e.plan)
-          case _ => Seq.empty
-        })
-      }.distinct
-    }
-
-    collectTempViews(child)
   }
 }
